Move data seeding to repository interface
The data seeding is moved to be default implementation in the repository interface
